Add CONFIG_SCHED_PDS in sched.h to ignore list
Had to do this to compile on Ubuntu due to error when generating header package
where the scripts complains about leaking CONFIG_SCHED_PDS to user-space.

This might be some test that Ubuntu does when generating header files for
package installation perhaps? There was no error during the compilation, and
setting this to "ignore" generated packages without errors.